+++ NEWS.md
@@ -2,6 +2,13 @@
 
 * `slicing()` with a 1-column matrix is now deprecated.
 
+* `row_number()`, `min_rank()`, `dense_rank()`, `ntile()`, `cume_dist()`, and
+  `percent_rank()` are now powered by vctrs, meaning that they are faster and
+  work for more types. You can now also rank by multiple columns at once by
+  supplying a data frame to these functions (#6428).
+  
+* `ntile()` now requires `n` to be a single positive integer.
+
 * `transmute()` is superseded in favour of `mutate(.keep = "none")`
 
 * `recode()` is superseded in favor of `case_match()`. `recode_factor()` is
@@ -324,6 +331,10 @@
     `"error"` if dropped rows would be surprising.
 
 * `nest_join()` has gained the `na_matches` argument that all other joins have.
+
+# dplyr 1.0.10
+
+Hot patch release to resolve R CMD check failures.
 
 # dplyr 1.0.9